backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
python-multipart
supabase
httpx
pymupdf
pillow
pillow-heif
